rhis-prism/requirements.txt
requests
firecrawl
weaviate-client
sentence-transformers
pydantic
pyyaml

